console: Refactor connect in gateway connection form
