+++ pkg/webui/locales/ja.json
@@ -549,6 +549,17 @@
   "console.containers.gateway-location-form.gateway-location-form.locationFromStatusMessage": "",
   "console.containers.gateway-location-form.gateway-location-form.setLocationManually": "",
   "console.containers.gateway-location-form.gateway-location-form.noLocationSetInfo": "",
+  "console.containers.gateway-location-form.index.updateLocationFromStatus": "",
+  "console.containers.gateway-location-form.index.updateLocationFromStatusDescription": "",
+  "console.containers.gateway-location-form.index.setGatewayLocation": "",
+  "console.containers.gateway-location-form.index.locationSource": "",
+  "console.containers.gateway-location-form.index.locationPrivacy": "",
+  "console.containers.gateway-location-form.index.placement": "",
+  "console.containers.gateway-location-form.index.indoor": "",
+  "console.containers.gateway-location-form.index.outdoor": "",
+  "console.containers.gateway-location-form.index.locationFromStatusMessage": "",
+  "console.containers.gateway-location-form.index.setLocationManually": "",
+  "console.containers.gateway-location-form.index.noLocationSetInfo": "",
   "console.containers.gateway-onboarding-form.gateway-provisioning-form.gateway-registration-form-section.index.requireAuthenticatedConnectionDescription": "{packetBrokerURL}など、他のネットワーク参加者が見ることができる情報を選択します",
   "console.containers.gateway-onboarding-form.gateway-provisioning-form.gateway-registration-form-section.index.shareGatewayInfoDescription": "ゲートウェイが{loraBasicStationURL}で駆動している場合など、このオプションを選択します",
   "console.containers.gateway-onboarding-form.gateway-provisioning-form.gateway-registration-form-section.validation-schema.validateEntry": "",
